Rewrite move ordering.
Moves are now scored by 16-bit integers. Create a ScoredMove struct to hold moves and scores. Even captures are scored above all quiet moves. Delay SEE on captures until the move is selected by MVV/LVA.
bench: 2715181
791-638-1601 SPRT [0, 5] at 10+0.1 (17.56 +/- 8.49)
388-255-1177 SPRT [0, 5] at 60+0.6 (25.43 +/- 9.47)

+++ src/common.h
@@ -184,6 +184,6 @@ public:
 };
 
 typedef SearchArrayList<Move> MoveList;
-typedef SearchArrayList<int> ScoreList;
+typedef SearchArrayList<int16_t> ScoreList;
 
 #endif
